feat: feed comment

+++ internal/entity/doc.go
@@ -48,7 +48,7 @@ type SearchDocsQuery struct {
 }
 
 type GetDocsResp struct {
-	Total int       `json:"total"`
+	Total int64     `json:"total"`
 	List  []DocInfo `json:"list"`
 }
